+++ CHANGES.md
@@ -1,4 +1,4 @@
-Version 1.1.0, 2014.??.??
+## Version 1.1.0, 2014.09.17
 
 * [#2](https://github.com/neocotic/mor.js/issues/2): Add unit tests
 * Add [Bower][2] support
@@ -8,12 +8,12 @@ Version 1.1.0, 2014.??.??
 * Remove [Ender][4] support
 * Rewrite API
 
-Version 1.0.2, 2011.12.17
+## Version 1.0.2, 2011.12.17
 
 * [#1](https://github.com/neocotic/mor.js/issues/1): Add support for single string or object argument to the `encode` and `decode` methods
 * [#3](https://github.com/neocotic/mor.js/issues/3): Expose read-only versions of the internal `chars` and `modes` variables
 
-Version 1.0.1, 2011.11.10
+## Version 1.0.1, 2011.11.10
 
 * Add support for [node.js][0], [CommonJS][3] and [Ender][4]
 * Add `VERSION` property to the API